pyproject.toml
[project]
name = "marketmenow"
version = "0.1.0"
description = "Modular marketing automation framework with a ports-and-adapters architecture"
readme = "README.md"
authors = [
    { email = "arnrustagi@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "google-genai>=1.0",
    "elevenlabs>=1.3",
    "typer[all]>=0.12",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.scripts]
mmn = "adapters.instagram.cli:app"

[tool.uv.build-backend]
module-name = ["marketmenow", "adapters"]

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"
